Fix QUIET mode display issue
Fixed issue where the temporary display wake in QUIET mode was not working. This was related to the timer being reset when a new image is drawn. Also fixed issue where the display energy saving mode (part of QUIET mode) would not enable after midnight. Bit of a hacky solution but good enough I think.